Refactor NEE++ into RenderPass

+++ src/Renderer/CPUDataStructures/NEEPlusPlusCPUData.h
@@ -17,8 +17,8 @@ struct NEEPlusPlusCPUData : public NEEPlusPlusCPUGPUCommonData
 	int frame_timer_before_visibility_map_update = 32;
 
 	std::vector<AtomicType<unsigned int>> packed_buffer;
-	AtomicType<unsigned int> total_shadow_ray_queries;
-	AtomicType<unsigned int> shadow_rays_actually_traced;
+	AtomicType<unsigned long long int> total_shadow_ray_queries;
+	AtomicType<unsigned long long  int> shadow_rays_actually_traced;
 };
 
 #endif
